+++ rs-react-app/package.json
@@ -17,8 +17,8 @@
     "@json2csv/plainjs": "^7.0.6",
     "@reduxjs/toolkit": "^2.5.1",
     "next": "^15.2.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "react-redux": "^9.2.0",
     "redux": "^5.0.1"
   },
@@ -37,6 +37,7 @@
     "@vitejs/plugin-react": "^4.3.4",
     "babel-jest": "^29.7.0",
     "eslint": "^9.19.0",
+    "eslint-config-next": "^15.2.1",
     "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-react": "^7.37.4",
