+++ rs-react-app/src/app/page.tsx
@@ -1,5 +1,7 @@
+"use client"
+
 import { useEffect } from 'react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
 const Home = () => {
   const router = useRouter();
