+++ rs-react-app/src/types/index.ts
@@ -1,0 +1,4 @@
+export type QueryParams = {
+    pageNumber: string;
+    details: string;
+};
